desmond-hw4-airflow/requirements.txt
pandas
scikit-learn
boto3
apache-airflow
numpy
requests
beautifulsoup4
xgboost
pyspark
matplotlib
apache-airflow
apache-airflow-providers-amazon

